Move setting initial position out of filters.
