+++ assets/src/components/index.js
@@ -12,7 +12,6 @@ export { default as withAnimationControls } from './with-animation-controls';
 export { default as withAttributes } from './with-attributes';
 export { default as withBlockName } from './with-block-name';
 export { default as withHasSelectedInnerBlock } from './with-has-selected-inner-block';
-export { default as withInitialPositioning } from './with-initial-positioning';
 export { default as withPageNumber } from './with-page-number';
 export { default as withParentBlock } from './with-parent-block';
 export { default as withIsReordering } from './with-is-reordering';
